Refactor sync progress handling into a dedicated module
- Moved sync progress-related types and functionality from `commands_sync.rs` to a new `progress.rs` file.
- Introduced `SyncProgressPhase` and `SyncProgressUpdate` structs to manage sync state and updates.
- Added `SyncObserver` trait for observing sync progress updates.
- Implemented `emit_progress` function to facilitate progress reporting through the observer pattern.
